requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pandas==2.1.1
numpy==1.26.0
pydantic==2.4.2
google-generativeai==0.3.1
openai==1.13.3
python-multipart==0.0.6
tenacity==8.2.3
openpyxl==3.1.2
websockets==11.0.3
python-dotenv==1.0.0